Update to the about page
The Why and What sections of this page were talking about internal reasons for this version of Fuego, something that 99.9999% of the people who will read this page would have no context of or care about. The fact Player-UI was rewritten in Typescript is not something relevant to people exploring this repo. This page should talk about what the customer problem the framework is solving. What is Player-UI, what problem does it solve, and why should I use it rather than using React, or Angular, etc, etc.

I'm super excited that you guys got this open sources finally, but the docs need a lot more context and examples if you want to see any adoption of it.